Initial commit
In the initial commit, we got all the gems added to the Gemfile and configured that we like to use. At this point, this is a barebones template. This does not include gems which may add assets and application functionality, but rather the basis to start working from. We excluded things using .gitignore to ensure that various files didn't get committed to the repo.